+++ src/constants/url.js
@@ -1,1 +1,2 @@
 export const GET_ALL_NURSES_URL = process.env.serverUrl + "/nurses";
+export const CREATE_NEW_NURSE_URL = process.env.serverUrl + "/nurses"